gltf加载器现在有限支持多个primitive的加载 可以加载（修正过的）科拉莉的模型了 索引的类型改为了uint32_t，因该还有更好的变法

+++ sources/runtime/src/function/renderer/HardcodeAssets.h
@@ -2,6 +2,7 @@
 
 #include "core/cgmath.h"
 #include "platform/graphics/Mesh.h"
+#include <cstdint>
 #include <vector>
 
 // 硬编码的一些数据
@@ -14,7 +15,7 @@ const std::vector<Vector3f> skybox_cube_vertices = {{-1.0, -1.0, -1.0}, {1.0, -1
 const VertexLayout skybox_cube_vertex_layout{{{VertexAttribute::POSITION, Meta::FieldType::vec3f, 0}},
                                              sizeof(Vector3f)};
 
-const std::vector<uint16_t> skybox_cube_indices = {1, 0, 3, 3, 2, 1, 3, 7, 6, 6, 2, 3, 7, 3, 0, 0, 4, 7,
+const std::vector<uint32_t> skybox_cube_indices = {1, 0, 3, 3, 2, 1, 3, 7, 6, 6, 2, 3, 7, 3, 0, 0, 4, 7,
                                                    2, 6, 5, 5, 1, 2, 4, 5, 6, 6, 7, 4, 5, 4, 0, 0, 1, 5};
 
 struct Vertex {
@@ -38,5 +39,5 @@ const VertexLayout plane_vertices_vertex_layout{
      {VertexAttribute::UV, Meta::FieldType::vec2f, offsetof(Vertex, uv)}},
     sizeof(Vertex)};
 
-const std::vector<uint16_t> plane_indices = {0, 1, 2, 2, 3, 0};
+const std::vector<uint32_t> plane_indices = {0, 1, 2, 2, 3, 0};
 } // namespace Goonya::Graphics::Assets
